make catalog and /collections customizable
